+++ packages/agtree/src/index.ts
@@ -143,6 +143,7 @@ export {
     REMOVE_PROPERTY,
 } from './converter/data/css';
 
+// FIXME: check if needed
 // CSS utilities
 export { CssTokenStream } from './parser/css/css-token-stream';